+++ src/client/components/evaluate/detail/EqualPopulation.tsx
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { Box, Flex, jsx, Styled, ThemeUIStyleObject, Heading } from "theme-ui";
 import { EvaluateMetricWithValue, DistrictProperties } from "../../../../shared/entities";
-import { getChoroplethStops } from "../../map/index";
+import { getEqualPopulationStops } from "../../map/index";
 import { DistrictsGeoJSON } from "../../../types";
 
 const style: ThemeUIStyleObject = {
@@ -66,19 +66,21 @@ const EqualPopulationMetricDetail = ({
   readonly geojson?: DistrictsGeoJSON;
 }) => {
   function computeRowFill(row: DistrictProperties) {
-    const val = row.percentDeviation;
-    const choroplethStops = getChoroplethStops(metric.key, metric.popThreshold);
-    // eslint-disable-next-line
-    for (let i = 0; i < choroplethStops.length; i++) {
-      const r = choroplethStops[i];
-      if (val && val >= r[0]) {
-        if (i < choroplethStops.length - 1) {
-          const r1 = choroplethStops[i + 1];
-          if (val < r1[0]) {
+    const val = row.populationDeviation;
+    if (metric.popThreshold !== undefined && metric.avgPopulation !== undefined) {
+      const choroplethStops = getEqualPopulationStops(metric.popThreshold, metric.avgPopulation);
+      // eslint-disable-next-line
+      for (let i = 0; i < choroplethStops.length; i++) {
+        const r = choroplethStops[i];
+        if (val !== undefined && val >= r[0]) {
+          if (i < choroplethStops.length - 1) {
+            const r1 = choroplethStops[i + 1];
+            if (val < r1[0]) {
+              return r[1];
+            }
+          } else {
             return r[1];
           }
-        } else {
-          return r[1];
         }
       }
     }
@@ -110,7 +112,8 @@ const EqualPopulationMetricDetail = ({
                   <Styled.td sx={{ ...style.td, ...style.colFirst }}>{id}</Styled.td>
 
                   <Styled.td sx={style.td}>
-                    {feature.properties.percentDeviation ? (
+                    {feature.properties.percentDeviation !== undefined &&
+                    feature.properties.populationDeviation !== undefined ? (
                       <Flex sx={{ alignItems: "center" }}>
                         <Styled.div
                           sx={{
@@ -129,8 +132,13 @@ const EqualPopulationMetricDetail = ({
                   </Styled.td>
 
                   <Styled.td sx={{ ...style.td, ...style.number, ...style.colLast }}>
-                    {feature.properties.populationDeviation ? (
-                      Math.round(feature.properties.populationDeviation).toLocaleString()
+                    {feature.properties.populationDeviation !== undefined ? (
+                      Math.ceil(feature.properties.populationDeviation) === 0 ? (
+                        // Need this to handle negative 0, which Math.ceil likes to return
+                        "0"
+                      ) : (
+                        Math.ceil(feature.properties.populationDeviation).toLocaleString()
+                      )
                     ) : (
                       <Box sx={style.blankValue}>-</Box>
                     )}